refator(map markers): move all marker junk into own directory
