+++ packages/vehicles/src/VehicleMarker/RotatedMarker.js
