+++ packages/vehicles/src/VehicleMarker/icons.js
